[MBFYCNCT-45] Create product/utils.js file
